mod: rename models/hyperdrive to models/registry
Internal directory rename only. External SDK exports remain unchanged.
