+++ packages/qvac-sdk/server/rpc/handlers/load-model/hyperdrive.ts
@@ -23,7 +23,7 @@ import {
   extractTensorsFromShards,
   calculatePercentage,
 } from "@/server/utils";
-import { getModelBySrc } from "@/models/hyperdrive";
+import { getModelBySrc } from "@/models/registry";
 import {
   getActiveDownload,
   registerDownload,
